# backend/public/swagger.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Server"
    },
    {
      "url": "https://blog-api.biud436.com",
      "description": "Production Server"
    }
  ],
  "info": {
    "version": "1.0.0",
    "title": "blog-api-server",
    "description": "블로그 API 서버입니다."
  },
  "paths": {
...
    "/admin/temp/post": {
      "post": {
        "tags": [
          "관리"
        ],
        "parameters": [],
        "requestBody": {
          "description": "임시 포스트 정보",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostTempDto"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResponsableData"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "임시 포스트를 저장합니다",
        "description": "임시 포스트를 저장합니다.",
        "x-nestia-namespace": "admin.temp.post.saveTempPost",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "관리",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "userId",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "유저 ID",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "createPostTempDto",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "임시 포스트 정보",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ]
      },
      "get": {
        "tags": [
          "관리"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResponsableData"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "특정 유저에 대한 모든 임시 포스트를 조회합니다",
        "description": "특정 유저에 대한 모든 임시 포스트를 조회합니다.",
        "x-nestia-namespace": "admin.temp.post.getTempAllPost",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "관리",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "userId",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "유저 ID",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ]
      }
    },
    "/admin/temp/post/{postId}": {
      "patch": {
        "tags": [
          "관리"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "포스트 번호",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "임시 포스트 정보",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostTempDto"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResponsableData"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "임시 포스트를 수정합니다",
        "description": "임시 포스트를 수정합니다.",
        "x-nestia-namespace": "admin.temp.post.updateTempPost",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "관리",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "userId",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "유저 ID",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "postId",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "포스트 번호",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "updatePostTempDto",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "임시 포스트 정보",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ]
      },
      "delete": {
        "tags": [
          "관리"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "포스트 번호",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResponsableData"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "임시 포스트를 삭제합니다",
        "description": "임시 포스트를 삭제합니다.",
        "x-nestia-namespace": "admin.temp.post.deleteTempPostById",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "관리",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "userId",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "유저 ID",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "postId",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "포스트 번호",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ]
      },
      "get": {
        "tags": [
          "관리"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "포스트 번호",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResponsableData"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "임시 포스트를 조회합니다",
        "description": "임시 포스트를 조회합니다.",
        "x-nestia-namespace": "admin.temp.post.getTempPostById",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "관리",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "userId",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "유저 ID",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "postId",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "포스트 번호",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ]
      }
    },
...
    "/api/post/{postId}": {
      "delete": {
        "tags": [
          "API"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResponsableData"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "Delete post by post id in the admin page",
        "description": "Delete post by post id in the admin page.",
        "x-nestia-namespace": "api.post.deletePost",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "API",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "postId",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ]
      }
    },
    "/api/post": {
      "get": {
        "tags": [
          "API"
        ],
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "description": "",
            "schema": {
              "type": "number"
            },
            "required": true
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResponsableData"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "Gets post list in the admin page",
        "description": "Gets post list in the admin page.",
        "x-nestia-namespace": "api.post.getPost",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "API",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "pageNumber",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "pageSize",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ]
      }
    },
    "/post/breadcrumbs": {
      "get": {
        "tags": [
          "Post"
        ],
        "parameters": [
          {
            "name": "categoryName",
            "in": "query",
            "description": "카테고리 이름",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResponsableData"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "breadcrumbs 정보를 조회합니다",
        "description": "breadcrumbs 정보를 조회합니다.",
        "x-nestia-namespace": "post.breadcrumbs.getBreadcrumbs",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "Post",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "categoryName",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "카테고리 이름",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ]
      }
    },
    "/posts/breadcrumbs": {
      "get": {
        "tags": [
          "Post"
        ],
        "parameters": [
          {
            "name": "categoryName",
            "in": "query",
            "description": "카테고리 이름",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResponsableData"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "breadcrumbs 정보를 조회합니다",
        "description": "breadcrumbs 정보를 조회합니다.",
        "x-nestia-namespace": "posts.breadcrumbs.getBreadcrumbs",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "Post",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "categoryName",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "카테고리 이름",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ]
      }
    },
...
    "/post": {
      "post": {
        "tags": [
          "Post"
        ],
        "parameters": [],
        "requestBody": {
          "description": "생성할 포스트 정보",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResponsableData"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "description": "새로운 포스트를 생성합니다 (관리자만 가능)",
        "x-nestia-namespace": "post.create",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "Post",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "userId",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "JWT에서 추출한 유저 아이디",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "createPostDto",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "생성할 포스트 정보",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ]
      },
      "get": {
        "tags": [
          "Post"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "페이지 번호 (1부터 시작)",
            "schema": {
              "type": "number"
            },
            "required": true
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "카테고리 ID, 생략하면 전체 포스트를 조회합니다.",
            "schema": {
              "type": "number"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/IResponsableData"
                    },
                    {
                      "$ref": "#/components/schemas/ResponseUtil.FailureResponse"
                    }
                  ]
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "포스트를 페이지 단위로 조회합니다",
        "description": "포스트를 페이지 단위로 조회합니다.",
        "x-nestia-namespace": "post.findAll",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "Post",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "page",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "페이지 번호 (1부터 시작)",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "categoryId",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "카테고리 ID, 생략하면 전체 포스트를 조회합니다.",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ]
      }
    },
...
    "/post/{id}/comment": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResponsableData"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "댓글을 조회합니다",
        "description": "댓글을 조회합니다.",
        "x-nestia-namespace": "post.comment.getComments",
        "x-nestia-jsDocTags": [
          {
            "name": "param",
            "text": [
              {
                "text": "id",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "댓글 ID",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ]
      }
    },
    "/posts/{id}/comment": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResponsableData"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "댓글을 조회합니다",
        "description": "댓글을 조회합니다.",
        "x-nestia-namespace": "posts.comment.getComments",
        "x-nestia-jsDocTags": [
          {
            "name": "param",
            "text": [
              {
                "text": "id",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "댓글 ID",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ]
      }
    },
    "/post/comment": {
      "post": {
        "tags": [
          "Post"
        ],
        "parameters": [],
        "requestBody": {
          "description": "생성할 댓글 정보",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResponsableData"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "새로운 댓글을 작성합니다",
        "description": "새로운 댓글을 작성합니다.",
        "x-nestia-namespace": "post.comment.createComment",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "Post",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "userId",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "JWT에서 추출한 유저 아이디",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "createCommentDto",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "생성할 댓글 정보",
                "kind": "text"
              }
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePostTempDto": {
        "type": "object",
        "properties": {
          "title": {
            "title": "제목",
            "description": "@title 제목",
            "x-typia-jsDocTags": [
              {
                "name": "title",
                "text": [
                  {
                    "text": "제목",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string"
          },
          "content": {
            "title": "내용",
            "description": "@title 내용",
            "x-typia-jsDocTags": [
              {
                "name": "title",
                "text": [
                  {
                    "text": "내용",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string"
          }
        },
        "nullable": false,
        "required": [
          "title",
          "content"
        ],
        "description": "@title 내용",
        "x-typia-jsDocTags": []
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}